# lib/bytesutil: Optimize ToUnsafeBytes to make it leaner, more standards-compliant and slightly faster